Refactor: chatgpt code
