--- cogs/cog_chatGPT/agent/__init__.py
@@ -1,2 +1,0 @@
-from .autoGPT import AutoGPTAgent
-from .basic import Basic
