+++ cogs/cog_chatGPT/chat/tool/base.py
@@ -1,0 +1,7 @@
+from langchain_core.tools import BaseTool
+
+
+class ToolBase:
+    name: str
+    description: str
+    tool: BaseTool
